pyproject.toml
[project]
name = "seedream"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "lancedb>=0.26.0",
    "marked>=0.9.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "uvicorn[standard]>=0.40.0",
]
